change StatChart props so backgroundColor is passed from StatChartPanel
