+++ ui/components/src/StatChart.tsx
@@ -13,7 +13,6 @@
 
 import { useMemo } from 'react';
 import { merge } from 'lodash-es';
-import { useTheme } from '@mui/material';
 import type { EChartsOption } from 'echarts';
 import { use } from 'echarts/core';
 import { GaugeChart as EChartsGaugeChart, GaugeSeriesOption } from 'echarts/charts';
@@ -22,7 +21,6 @@ import { GridComponent, DatasetComponent, TitleComponent, TooltipComponent } fro
 import { CanvasRenderer } from 'echarts/renderers';
 import { GraphSeries } from '@perses-dev/plugin-system';
 import { formatValue, UnitOptions } from './model/units'; // TODO (sjcobb): add back formatValue
-import { defaultThresholdInput, ThresholdOptions } from './model/thresholds';
 import { EChartsWrapper } from './EChartsWrapper';
 
 use([
@@ -67,22 +65,12 @@ interface StatChartProps {
   height: number;
   data: StatChartData;
   unit: UnitOptions;
-  thresholds?: ThresholdOptions;
+  backgroundColor?: string;
   sparkline?: LineSeriesOption;
 }
 
 export function StatChart(props: StatChartProps) {
-  const { width, height, data, unit, sparkline } = props;
-  const thresholds = props.thresholds ?? defaultThresholdInput;
-  const theme = useTheme();
-
-  const showSparkline = sparkline !== undefined ? true : false;
-  let backgroundColor = 'transparent';
-  if (thresholds.default_color) {
-    backgroundColor = thresholds.default_color;
-  } else if (showSparkline === true) {
-    backgroundColor = theme.palette.primary.light;
-  }
+  const { width, height, data, unit, backgroundColor, sparkline } = props;
 
   const option: EChartsOption = useMemo(() => {
     if (data.seriesData === undefined) return {};
@@ -91,7 +79,6 @@ export function StatChart(props: StatChartProps) {
     const series = data.seriesData;
     const calculatedValue = data.calculatedValue ?? 0;
     const isLargePanel = width > 250 ? true : false;
-    //  const nameFontSize = isLargePanel ? 30 : 12;
     const showName = isLargePanel;
     const name = showName === true ? data.name : '';
     const smallestSide = Math.min(width, height * 1.2);
