+++ ui/panels-plugin/src/plugins/stat-chart/StatChartPanel.tsx
@@ -14,6 +14,7 @@
 import { JsonObject } from '@perses-dev/core';
 import { StatChart, StatChartData, defaultThresholdInput, ThresholdOptions } from '@perses-dev/components';
 import { Box, Skeleton } from '@mui/material';
+import { useTheme } from '@mui/material';
 import { LineSeriesOption } from 'echarts/charts';
 import { useMemo } from 'react';
 import { GraphQueryDefinition, GraphData, useGraphQuery, PanelProps } from '@perses-dev/plugin-system';
@@ -50,10 +51,19 @@ export function StatChartPanel(props: StatChartPanelProps) {
     },
     contentDimensions,
   } = props;
-  const thresholds = props.definition.options.thresholds ?? defaultThresholdInput;
   const suggestedStepMs = useSuggestedStepMs(contentDimensions?.width);
   const { data, loading, error } = useGraphQuery(query, { suggestedStepMs });
   const chartData = useChartData(data, calculation, name);
+  const theme = useTheme();
+
+  const thresholds = props.definition.options.thresholds ?? defaultThresholdInput;
+  const showSparkline = sparkline !== undefined ? true : false;
+  let backgroundColor = 'transparent';
+  if (thresholds.default_color) {
+    backgroundColor = thresholds.default_color;
+  } else if (showSparkline === true) {
+    backgroundColor = theme.palette.primary.light;
+  }
 
   if (error) throw error;
 
@@ -77,7 +87,7 @@ export function StatChartPanel(props: StatChartPanelProps) {
       height={contentDimensions.height}
       data={chartData}
       unit={unit}
-      thresholds={thresholds}
+      backgroundColor={backgroundColor}
       sparkline={convertSparkline(sparkline)}
     />
   );
